requirements.txt
apache-airflow
psycopg2-binary
python-dotenv
SQLAlchemy